# specs/001-fundamentos-identidad/contracts/openapi-users-service.yaml
openapi: 3.0.3
paths:
  /health/live:
    get:
      operationId: usersLiveness
      summary: ""
      parameters: []
      responses:
        "200":
          description: Process is alive.
  /health/ready:
    get:
      operationId: usersReadiness
      summary: ""
      parameters: []
      responses:
        "200":
          description: Database is ready.
        "503":
          description: Database is unavailable.
  /internal/v1/registrations:
    post:
      operationId: createPendingUser
      summary: ""
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - registrationId
                - userId
                - name
                - email
                - role
              properties:
                registrationId:
                  type: string
                  format: uuid
                userId:
                  type: string
                  format: uuid
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                email:
                  type: string
                  format: email
                  maxLength: 254
                role:
                  type: string
                  enum:
                    - GUEST
                    - OWNER
      responses:
        "201":
          description: Pending user created or same command replayed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  email:
                    type: string
                    format: email
                  role:
                    type: string
                    enum:
                      - GUEST
                      - OWNER
                      - ADMIN
                  status:
                    type: string
                    enum:
                      - PENDING
                      - ACTIVE
                      - CANCELLED
        "400":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a1
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a1
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a1
        "409":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a1
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a1
      security: &a3
        - serviceAuth: []
  /internal/v1/registrations/{registrationId}/activate:
    post:
      operationId: activatePendingUser
      summary: ""
      parameters:
        - name: registrationId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: User ACTIVE; replay is idempotent.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - name
                  - email
                  - role
                  - status
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  email:
                    type: string
                    format: email
                  role:
                    type: string
                    enum:
                      - GUEST
                      - OWNER
                      - ADMIN
                  status:
                    type: string
                    enum:
                      - PENDING
                      - ACTIVE
                      - CANCELLED
        "400":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a2
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a2
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a2
        "404":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a2
        "409":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a2
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a2
      security: *a3
  /internal/v1/registrations/{registrationId}/cancel:
    post:
      operationId: cancelPendingUser
      summary: ""
      parameters:
        - name: registrationId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Pending user cancelled; replay is idempotent.
        "400":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a4
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a4
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a4
        "404":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a4
        "409":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a4
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a4
      security: *a3
  /internal/v1/login-identities/resolve:
    post:
      operationId: resolveLoginIdentity
      summary: ""
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
      responses:
        "200":
          description: Minimal ACTIVE identity.
          content:
            application/json:
              schema:
                type: object
                required:
                  - userId
                  - role
                  - status
                properties:
                  userId:
                    type: string
                    format: uuid
                  role:
                    type: string
                    enum:
                      - GUEST
                      - OWNER
                      - ADMIN
                  status:
                    type: string
                    enum:
                      - ACTIVE
        "400":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a5
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a5
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a5
        "404":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a5
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a5
      security:
        - serviceAuth: []
  /internal/v1/users/{userId}/profile:
    get:
      operationId: getProfile
      summary: ""
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Sanitized profile.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - name
                  - email
                  - role
                  - version
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  email:
                    type: string
                    format: email
                  role:
                    type: string
                    enum:
                      - GUEST
                      - OWNER
                      - ADMIN
                  phone:
                    type: string
                    nullable: true
                  preferences:
                    allOf:
                      - type: object
                        maxProperties: 20
                        additionalProperties:
                          oneOf:
                            - type: string
                            - type: number
                            - type: boolean
                    nullable: true
                  photoUrl:
                    type: string
                    nullable: true
                  version:
                    type: integer
                    minimum: 1
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a6
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a6
        "404":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a6
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a6
      security: &a9
        - bearerAuth: []
    patch:
      operationId: updateProfile
      summary: ""
      description: Atomic partial update. Omitted fields are preserved; null clears
        optional values. File plus photo:null and empty changes are invalid.
        Authentication precedes ownership and input validation.
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required:
                - profile
              properties:
                profile:
                  type: object
                  additionalProperties: false
                  required:
                    - expectedVersion
                  properties:
                    expectedVersion:
                      type: integer
                      minimum: 1
                    name:
                      type: string
                      minLength: 2
                      maxLength: 100
                    email:
                      type: string
                      format: email
                      maxLength: 254
                    phone:
                      type: string
                      nullable: true
                      pattern: ^\+[1-9][0-9]{7,14}$
                    preferences:
                      allOf:
                        - &a7
                          type: object
                          maxProperties: 20
                          additionalProperties:
                            oneOf:
                              - type: string
                              - type: number
                              - type: boolean
                      nullable: true
                    photo:
                      type: string
                      nullable: true
                      enum:
                        - null
                photo:
                  type: string
                  format: binary
                  description: JPEG or PNG, maximum 5,000,000 bytes (5 MB decimal).
            encoding:
              profile:
                contentType: application/json
              photo:
                contentType: image/jpeg, image/png
      responses:
        "200":
          description: Atomic update complete.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - name
                  - email
                  - role
                  - version
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  email:
                    type: string
                    format: email
                  role:
                    type: string
                    enum:
                      - GUEST
                      - OWNER
                      - ADMIN
                  phone:
                    type: string
                    nullable: true
                  preferences:
                    allOf:
                      - *a7
                    nullable: true
                  photoUrl:
                    type: string
                    nullable: true
                  version:
                    type: integer
                    minimum: 1
        "400":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a8
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
        "404":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
        "409":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
        "413":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
        "415":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a8
      security: *a9
  /internal/v1/users/{userId}/profile/photo:
    get:
      operationId: getProfilePhoto
      summary: ""
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: JPEG/PNG bytes.
          headers:
            ETag:
              schema:
                type: string
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        "401":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: &a10
                type: object
                required:
                  - type
                  - title
                  - status
                  - detail
                  - instance
                  - code
                  - traceId
                  - errors
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  code:
                    type: string
                  traceId:
                    type: string
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - field
                        - code
                      properties:
                        field:
                          type: string
                        code:
                          type: string
        "403":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a10
        "404":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a10
        "503":
          description: Safe Problem Details response.
          content:
            application/problem+json:
              schema: *a10
      security: *a9
info:
  title: StayHub Users Service API
  description: Internal REST API; users-service owns identity, role and profile data.
  version: 1.0.0
  contact: {}
tags: []
servers:
  - url: http://users-service:3002
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
    serviceAuth:
      scheme: bearer
      bearerFormat: Service-JWT
      type: http
  schemas: {}